refactor: `model` 폴더 생성
