+++ src/main/java/lotto/model/Result.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.model;
 
 import java.util.Objects;
